Move artifacts management into java part
